web/requirements.txt
django
gunicorn
djangorestframework
requests
dotenv
celery==5.2.7
redis==4.4.0
